Modal implementation (#12)
* adding basic file structure and linking Bulma CSS framework and jQuery/jQueryUI

* API can be successfully called with a hard coded search; first 10 results are displayed in tiles on the page

* demo data in html while working on click functions; now clicking on a recipe tile passes the recipe id to a function that will make an api call to get recipe information; next step - create modal to pop up with recipe info, option to save recipe

* added CSS line to make cursor over tiles be a pointer to indicate they can be clicked; changed dummy data to remove anchor from title to match change in javascript; set up API headers for more detailed recipe information for the selected recipe tile

* Added basic modal

A card modal is pulled up when a recipe icon is clicked, and will close when close is clicked. Recipe information or save functionality is not yet added to the modal.

* Added more modal functionality

Modal now *attempts* to print data from the API, not yet fully functional. Also added a second close modal function, one for close and one for the x at top. Added section for add to list and forward/back buttons

* Improvements

Moved api population for modal to within the api call. began switching it to use jquery. added beginning code for next/back buttons.

* Modal implementation

Not perfect buy printing data from api onto modal functions

* added modal clear function

function to clear out the modal

Co-authored-by: Kathleen P <30782087+kpehl@users.noreply.github.com>